ai_model/requirements.txt
pdfminer3
pyresparser
streamlit
pandas
pafy
plotly
pymysql
streamlit-tags
Pillow
youtube-dl
nltk
pdfminer3
spacy
pdfminer.six
google-generativeai
parrot
textblob
language_tool_python
tensorflow
keras
scikit-learn
joblib
pandas